requirements.txt
groq 
streamlit 
requests 
speechrecognition 
pyttsx3 
python-dotenv
PyAudio
streamlit_option_menu
